sysmexcbctools/data/requirements.txt
# Core dependencies
pandas>=1.2.0
numpy>=1.19.0
pyyaml>=5.4.0
tqdm>=4.60.0
psutil>=5.8.0

# Optional: Memory optimization
# dask[dataframe]>=2021.3.0
# pyarrow>=3.0.0

# Development dependencies
# pytest>=6.2.0
# pytest-cov>=2.12.0
